Initial port of OpenZFS to CHERI
This port supports CheriABI programs atop a hybrid kernel (and maybe
purecap, but that is untested.) There is no support for legacy ioctls or
for freebsd64 as much of the ioctl subsystem and libnvpair will need to
be rewritten to accomodate such a thing.

This port changes the uint64_t's in ioctl structures to be kuintcap64_t
which is a uint64_t unless in pure-capablity mode or in a hybrid kernel.

Casts of pointers to or through uint64_t's have been changed to use
uintptr_t or uintcap_t as appropriate.  Code that expects addresses
uses ptraddr_t.

Co-authored-by: Allison Randal <allison@lohutok.net>