freebsd64_sendmsg: allow more control messages
When a freebsd64 caller uses all or most allowed space for control
messages (MCLBYTES) then the message may no longer fit when the messages
are padded for capability alignment. Historically (with freebsd32) we've
just shrugged and said there is no ABI guarantee. Unfortunatly, libnv
breaks this assumption so capsicumized programs link nm fail on CHERI
systems.

Fix this by using external storage in mbufs if the expansion requires
it.